chore: rename folder customUi to custom UI and move paginationcontrol from src/components to src/components/customUi
