+++ src/components/OnboardingForm.tsx
@@ -11,7 +11,7 @@ import {
   TooltipTrigger,
 } from "@/components/ui/tooltip";
 import { Input } from "@/components/ui/input";
-import { PasswordInput } from "@/components/customUi/password-input";
+import { PasswordInput } from "@/components/customUI/password-input";
 import { Label } from "@/components/ui/label";
 import { useMutation } from "@tanstack/react-query";
 import { saveOnboardingInfo } from "@/lib/request/saveOnboardingInfo";
